requirements.txt
streamlit>=1.35.0
google-genai>=0.3.0
python-dotenv>=1.0.0
filelock>=3.12.0